servicesAPI/requirements.txt
flask >= 2.2.3
openai >= 0.27.2
python-dotenv >= 0.21.1
flask_cors>=3.0.10
